refactor(media): extract provider switch to its own module and run it async
Moves the provider-aware refresh decision, candidate finder, and episode
reconciliation out of the oversized media.ex into Mydia.Media.ProviderSwitch
(shared title/episode helpers exposed as @doc-false). The LiveView refresh
now runs the re-identify search and the provider switch via start_async so
the socket is never blocked during the multi-request reconciliation.
