+++ lib/mydia_web/live/media_live/show.ex
@@ -82,6 +82,7 @@ defmodule MydiaWeb.MediaLive.Show do
      |> assign(:show_reidentify_modal, false)
      |> assign(:reidentify_candidates, [])
      |> assign(:reidentify_provider, nil)
+     |> assign(:reidentifying, false)
      # Manual search modal state
      |> assign(:show_manual_search_modal, false)
      |> assign(:manual_search_query, "")
@@ -521,6 +522,12 @@ defmodule MydiaWeb.MediaLive.Show do
 
   def handle_async(:refresh_files, result, socket),
     do: FileEvents.handle_refresh_files_async(result, socket)
+
+  def handle_async(:reidentify_search, result, socket),
+    do: FileEvents.handle_reidentify_search_async(result, socket)
+
+  def handle_async(:reidentify_adopt, result, socket),
+    do: FileEvents.handle_reidentify_adopt_async(result, socket)
 
   def handle_async(:rescan_season_files, result, socket),
     do: FileEvents.handle_rescan_season_files_async(result, socket)
